api/requirements.txt
PyYAML==3.11
SQLAlchemy==1.0.13
Unirest==1.1.7
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcrypt==2.0.0
certifi==2016.2.28
cffi==1.6.0
dnspython==1.14.0
poster==0.8.1
psycopg2==2.7.4
pycparser==2.14
singledispatch==3.4.0.3
six==1.10.0
tornado==5.0.2
wsgiref==0.1.2
